[Output] Formatter handles string quoting
Had to refactor the handling of key comment to make this possible.

Very interesting is Xcode's escaping of newlines: It escapes all newlines except if the only newline is the last character. I will do the same here, as the whole idea is to mimic Xcode's output as closely as possible